simplify code by extracting helper Fn, config and msg store into its seperate files
