+++ dkg/frost/frost_blame.go
@@ -9,6 +9,100 @@ import (
 	"github.com/pkg/errors"
 )
 
+func (fr *FROST) createAndBroadcastBlameOfInconsistentMessage(existingMessage, newMessage *dkg.SignedMessage) (bool, *dkg.ProtocolOutcome, error) {
+	fr.state.currentRound = Blame
+
+	existingMessageBytes, err := existingMessage.Encode()
+	if err != nil {
+		return false, nil, err
+	}
+	newMessageBytes, err := newMessage.Encode()
+	if err != nil {
+		return false, nil, err
+	}
+
+	msg := &ProtocolMsg{
+		Round: Blame,
+		BlameMessage: &BlameMessage{
+			Type:             InconsistentMessage,
+			TargetOperatorID: uint32(newMessage.Signer),
+			BlameData:        [][]byte{existingMessageBytes, newMessageBytes},
+			BlamerSessionSk:  fr.config.sessionSK.Bytes(),
+		},
+	}
+
+	signedMessage, err := fr.broadcastDKGMessage(msg)
+	if err != nil {
+		return false, nil, err
+	}
+
+	return true, &dkg.ProtocolOutcome{
+		BlameOutput: &dkg.BlameOutput{
+			Valid:        true,
+			BlameMessage: signedMessage,
+		},
+	}, nil
+}
+
+func (fr *FROST) createAndBroadcastBlameOfInvalidShare(culpritOID uint32) (bool, *dkg.ProtocolOutcome, error) {
+	fr.state.currentRound = Blame
+
+	round1Bytes, err := fr.state.msgs[Round1][culpritOID].Encode()
+	if err != nil {
+		return false, nil, err
+	}
+	msg := &ProtocolMsg{
+		Round: Blame,
+		BlameMessage: &BlameMessage{
+			Type:             InvalidShare,
+			TargetOperatorID: culpritOID,
+			BlameData:        [][]byte{round1Bytes},
+			BlamerSessionSk:  fr.config.sessionSK.Bytes(),
+		},
+	}
+	signedMessage, err := fr.broadcastDKGMessage(msg)
+	if err != nil {
+		return false, nil, err
+	}
+	return true, &dkg.ProtocolOutcome{
+		BlameOutput: &dkg.BlameOutput{
+			Valid:        true,
+			BlameMessage: signedMessage,
+		},
+	}, nil
+}
+
+func (fr *FROST) createAndBroadcastBlameOfInvalidMessage(culpritOID uint32, message *dkg.SignedMessage) (bool, *dkg.ProtocolOutcome, error) {
+	fr.state.currentRound = Blame
+
+	bytes, err := message.Encode()
+	if err != nil {
+		return false, nil, err
+	}
+
+	msg := &ProtocolMsg{
+		Round: Blame,
+		BlameMessage: &BlameMessage{
+			Type:             InvalidMessage,
+			TargetOperatorID: culpritOID,
+			BlameData:        [][]byte{bytes},
+			BlamerSessionSk:  fr.config.sessionSK.Bytes(),
+		},
+	}
+
+	signedMsg, err := fr.broadcastDKGMessage(msg)
+	if err != nil {
+		return false, nil, err
+	}
+
+	return true, &dkg.ProtocolOutcome{
+		BlameOutput: &dkg.BlameOutput{
+			Valid:        true,
+			BlameMessage: signedMsg,
+		},
+	}, nil
+}
+
 func (fr *FROST) checkBlame(blamerOID uint32, protocolMessage *ProtocolMsg, signedMessage *dkg.SignedMessage) (finished bool, protocolOutcome *dkg.ProtocolOutcome, err error) {
 	fr.state.currentRound = Blame
 
@@ -144,124 +238,6 @@ func (fr *FROST) processBlameTypeInvalidMessage(blameMessage *BlameMessage) (boo
 		return true, nil
 	}
 	return false, errors.New("message is valid")
-}
-
-func (fr *FROST) decodeMessage(data []byte) (*dkg.SignedMessage, *ProtocolMsg, error) {
-	signedMsg := &dkg.SignedMessage{}
-	if err := signedMsg.Decode(data); err != nil {
-		return nil, nil, errors.Wrap(err, "failed to decode signed message")
-	}
-	pMsg := &ProtocolMsg{}
-	if err := pMsg.Decode(signedMsg.Message.Data); err != nil {
-		return signedMsg, nil, errors.Wrap(err, "failed to decode protocol msg")
-	}
-	return signedMsg, pMsg, nil
-}
-
-func (fr *FROST) createAndBroadcastBlameOfInconsistentMessage(existingMessage, newMessage *dkg.SignedMessage) (bool, *dkg.ProtocolOutcome, error) {
-	fr.state.currentRound = Blame
-
-	existingMessageBytes, err := existingMessage.Encode()
-	if err != nil {
-		return false, nil, err
-	}
-	newMessageBytes, err := newMessage.Encode()
-	if err != nil {
-		return false, nil, err
-	}
-
-	msg := &ProtocolMsg{
-		Round: Blame,
-		BlameMessage: &BlameMessage{
-			Type:             InconsistentMessage,
-			TargetOperatorID: uint32(newMessage.Signer),
-			BlameData:        [][]byte{existingMessageBytes, newMessageBytes},
-			BlamerSessionSk:  fr.config.sessionSK.Bytes(),
-		},
-	}
-
-	signedMessage, err := fr.broadcastDKGMessage(msg)
-	if err != nil {
-		return false, nil, err
-	}
-
-	return true, &dkg.ProtocolOutcome{
-		BlameOutput: &dkg.BlameOutput{
-			Valid:        true,
-			BlameMessage: signedMessage,
-		},
-	}, nil
-}
-
-func (fr *FROST) createAndBroadcastBlameOfInvalidShare(culpritOID uint32) (bool, *dkg.ProtocolOutcome, error) {
-	fr.state.currentRound = Blame
-
-	round1Bytes, err := fr.state.msgs[Round1][culpritOID].Encode()
-	if err != nil {
-		return false, nil, err
-	}
-	msg := &ProtocolMsg{
-		Round: Blame,
-		BlameMessage: &BlameMessage{
-			Type:             InvalidShare,
-			TargetOperatorID: culpritOID,
-			BlameData:        [][]byte{round1Bytes},
-			BlamerSessionSk:  fr.config.sessionSK.Bytes(),
-		},
-	}
-	signedMessage, err := fr.broadcastDKGMessage(msg)
-	if err != nil {
-		return false, nil, err
-	}
-	return true, &dkg.ProtocolOutcome{
-		BlameOutput: &dkg.BlameOutput{
-			Valid:        true,
-			BlameMessage: signedMessage,
-		},
-	}, nil
-}
-
-func (fr *FROST) createAndBroadcastBlameOfInvalidMessage(culpritOID uint32, message *dkg.SignedMessage) (bool, *dkg.ProtocolOutcome, error) {
-	fr.state.currentRound = Blame
-
-	bytes, err := message.Encode()
-	if err != nil {
-		return false, nil, err
-	}
-
-	msg := &ProtocolMsg{
-		Round: Blame,
-		BlameMessage: &BlameMessage{
-			Type:             InvalidMessage,
-			TargetOperatorID: culpritOID,
-			BlameData:        [][]byte{bytes},
-			BlamerSessionSk:  fr.config.sessionSK.Bytes(),
-		},
-	}
-
-	signedMsg, err := fr.broadcastDKGMessage(msg)
-	if err != nil {
-		return false, nil, err
-	}
-
-	return true, &dkg.ProtocolOutcome{
-		BlameOutput: &dkg.BlameOutput{
-			Valid:        true,
-			BlameMessage: signedMsg,
-		},
-	}, nil
-}
-
-func (fr *FROST) haveSameRoot(existingMessage, newMessage *dkg.SignedMessage) bool {
-	r1, err := existingMessage.GetRoot()
-	if err != nil {
-		return false
-	}
-	r2, err := newMessage.GetRoot()
-	if err != nil {
-		return false
-	}
-	return bytes.Equal(r1, r2)
 }
 
 type ErrBlame struct {
